c2-server/cli/requirements.txt
aiohttp==3.9.1
rich==13.7.0
websockets==12.0
Pillow==10.1.0
